+++ hydra-node/src/Hydra/API/ClientInput.hs
@@ -4,7 +4,7 @@ module Hydra.API.ClientInput where
 
 import Hydra.Prelude
 
-import Hydra.Tx (IsTx, TxIdType)
+import Hydra.Tx (ConfirmedSnapshot, IsTx (..), TxIdType)
 import Test.QuickCheck.Arbitrary.ADT (ToADTArbitrary)
 
 data ClientInput tx
@@ -16,6 +16,9 @@ data ClientInput tx
   | Close
   | Contest
   | Fanout
+  | -- XXX: This is deliberately outside of the protocol as it is not part of the normal protocol flow.
+    -- WARN: Given its consequences to the local state, it is a very risky thing to use.
+    SideLoadSnapshot {snapshot :: ConfirmedSnapshot tx}
   deriving stock (Generic)
 
 deriving stock instance IsTx tx => Eq (ClientInput tx)
@@ -23,7 +26,7 @@ deriving stock instance IsTx tx => Show (ClientInput tx)
 deriving anyclass instance IsTx tx => ToJSON (ClientInput tx)
 deriving anyclass instance IsTx tx => FromJSON (ClientInput tx)
 
-instance (Arbitrary tx, Arbitrary (TxIdType tx)) => Arbitrary (ClientInput tx) where
+instance (Arbitrary tx, Arbitrary (TxIdType tx), Arbitrary (UTxOType tx), IsTx tx) => Arbitrary (ClientInput tx) where
   arbitrary = genericArbitrary
 
   -- NOTE: Somehow, can't use 'genericShrink' here as GHC is complaining about
@@ -38,5 +41,6 @@ instance (Arbitrary tx, Arbitrary (TxIdType tx)) => Arbitrary (ClientInput tx) w
     Close -> []
     Contest -> []
     Fanout -> []
+    SideLoadSnapshot sn -> SideLoadSnapshot <$> shrink sn
 
-instance (Arbitrary tx, Arbitrary (TxIdType tx)) => ToADTArbitrary (ClientInput tx)
+instance (Arbitrary tx, Arbitrary (TxIdType tx), Arbitrary (UTxOType tx), IsTx tx) => ToADTArbitrary (ClientInput tx)