+++ hydra-node/test/Hydra/API/HTTPServerSpec.hs
@@ -9,7 +9,14 @@ import Data.Aeson (Result (Error, Success), eitherDecode, encode, fromJSON)
 import Data.Aeson qualified as Aeson
 import Data.Aeson.Lens (key, nth)
 import Data.Text qualified as Text
-import Hydra.API.HTTPServer (DraftCommitTxRequest (..), DraftCommitTxResponse (..), SubmitTxRequest (..), TransactionSubmitted, httpApp)
+import Hydra.API.HTTPServer (
+  DraftCommitTxRequest (..),
+  DraftCommitTxResponse (..),
+  SideLoadSnapshotRequest (..),
+  SubmitTxRequest (..),
+  TransactionSubmitted,
+  httpApp,
+ )
 import Hydra.API.ServerOutput (CommitInfo (CannotCommit, NormalCommit))
 import Hydra.API.ServerSpec (dummyChainHandle)
 import Hydra.Cardano.Api (
@@ -24,6 +31,7 @@ import Hydra.JSONSchema (SchemaSelector, prop_validateJSONSchema, validateJSON, 
 import Hydra.Ledger.Cardano (Tx)
 import Hydra.Ledger.Simple (SimpleTx)
 import Hydra.Logging (nullTracer)
+import Hydra.Tx (ConfirmedSnapshot (..))
 import Hydra.Tx.IsTx (UTxOType)
 import System.FilePath ((</>))
 import System.IO.Unsafe (unsafePerformIO)
@@ -50,6 +58,7 @@ spec = do
     roundtripAndGoldenSpecs (Proxy @(ReasonablySized (DraftCommitTxRequest Tx)))
     roundtripAndGoldenSpecs (Proxy @(ReasonablySized (SubmitTxRequest Tx)))
     roundtripAndGoldenSpecs (Proxy @(ReasonablySized TransactionSubmitted))
+    roundtripAndGoldenSpecs (Proxy @(ReasonablySized (SideLoadSnapshotRequest Tx)))
 
     prop "Validate /commit publish api schema" $
       prop_validateJSONSchema @(DraftCommitTxRequest Tx) "api.json" $
@@ -133,6 +142,14 @@ spec = do
           . key "subscribe"
           . key "message"
 
+    prop "Validate /snapshot publish api schema" $
+      prop_validateJSONSchema @(SideLoadSnapshotRequest Tx) "api.json" $
+        key "components" . key "messages" . key "SideLoadSnapshotRequest" . key "payload"
+
+    prop "Validate /snapshot subscribe api schema" $
+      prop_validateJSONSchema @(ConfirmedSnapshot Tx) "api.json" $
+        key "components" . key "schemas" . key "ConfirmedSnapshot"
+
     apiServerSpec
     describe "SubmitTxRequest accepted tx formats" $ do
       prop "accepts json encoded transaction" $
@@ -168,6 +185,7 @@ apiServerSpec = do
               cantCommit
               getNothing
               getNoSeenSnapshot
+              getNothing
               getPendingDeposits
               putClientInput
         )
@@ -199,12 +217,43 @@ apiServerSpec = do
               cantCommit
               getNothing
               getSeenSnapshot
+              getNothing
               getPendingDeposits
               putClientInput
           )
           $ do
             get "/snapshot/last-seen"
               `shouldRespondWith` 200{matchBody = matchJSON seenSnapshot}
+
+    describe "GET /snapshot" $ do
+      prop "responds correctly" $ \confirmedSnapshot -> do
+        let getConfirmedSnapshot = pure confirmedSnapshot
+        withApplication (httpApp @SimpleTx nullTracer dummyChainHandle testEnvironment defaultPParams cantCommit getNothing getNoSeenSnapshot getConfirmedSnapshot getPendingDeposits putClientInput) $ do
+          get "/snapshot"
+            `shouldRespondWith` case confirmedSnapshot of
+              Nothing -> 404
+              Just s -> 200{matchBody = matchJSON s}
+
+      prop "ok response matches schema" $ \(confirmedSnapshot :: ConfirmedSnapshot Tx) ->
+        withMaxSuccess 4
+          . withJsonSpecifications
+          $ \schemaDir -> do
+            let getConfirmedSnapshot = pure $ Just confirmedSnapshot
+            withApplication (httpApp @Tx nullTracer dummyChainHandle testEnvironment defaultPParams cantCommit getNothing getNoSeenSnapshot getConfirmedSnapshot getPendingDeposits putClientInput) $ do
+              get "/snapshot"
+                `shouldRespondWith` 200
+                  { matchBody =
+                      matchValidJSON
+                        (schemaDir </> "api.json")
+                        (key "channels" . key "/snapshot" . key "subscribe" . key "message" . key "payload")
+                  }
+
+    describe "POST /snapshot" $ do
+      prop "responds on valid requests" $ \(request :: SideLoadSnapshotRequest Tx) ->
+        withApplication (httpApp @Tx nullTracer dummyChainHandle testEnvironment defaultPParams cantCommit getNothing getNoSeenSnapshot getNothing getPendingDeposits putClientInput) $
+          do
+            post "/snapshot" (Aeson.encode request)
+            `shouldRespondWith` 200
 
     describe "GET /snapshot/utxo" $ do
       prop "responds correctly" $ \utxo -> do
@@ -218,6 +267,7 @@ apiServerSpec = do
               cantCommit
               getUTxO
               getNoSeenSnapshot
+              getNothing
               getPendingDeposits
               putClientInput
           )
@@ -243,6 +293,7 @@ apiServerSpec = do
                   cantCommit
                   getUTxO
                   getNoSeenSnapshot
+                  getNothing
                   getPendingDeposits
                   putClientInput
               )
@@ -268,6 +319,7 @@ apiServerSpec = do
                 cantCommit
                 getUTxO
                 getNoSeenSnapshot
+                getNothing
                 getPendingDeposits
                 putClientInput
             )
@@ -298,6 +350,7 @@ apiServerSpec = do
               getHeadId
               getNothing
               getNoSeenSnapshot
+              getNothing
               getPendingDeposits
               putClientInput
           )
@@ -334,6 +387,7 @@ apiServerSpec = do
                 getHeadId
                 getNothing
                 getNoSeenSnapshot
+                getNothing
                 getPendingDeposits
                 putClientInput
             )